app/requirements-server.txt
sanic-jwt
